requirements.txt
allure-pytest==2.11.0
allure-python-commons==2.11.0
pytest==7.1.3
pytest-forked==1.4.0
pytest-repeat==0.9.1
pytest-xdist==2.5.0
requests==2.28.1
selenium==4.5.0
webdriver-manager==3.8.3
